requirements.txt
sklearn
nltk
numpy
pandas
tqdm
python-docx
tensorflow-gpu>=2.0